% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.maxima.R
\name{identify.maxima}
\alias{identify.maxima}
\title{Identify maxima in a data frame}
\usage{
\method{identify}{maxima}(df, span = 11)
}
\arguments{
\item{df}{A \code{data.frame} generated by the \code{load.data} function.}

\item{threshold}{An integer defining threshold for local extrema identification}
}
\description{
Identify local extrema in a data frame
}
