% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.minima.R
\name{identify.minima}
\alias{identify.minima}
\title{Identify minima in a data frame}
\usage{
\method{identify}{minima}(df, span)
}
\arguments{
\item{df}{A \code{data.frame} generated by the \code{load.data} function.}

\item{threshold}{An integer defining threshold for local extrema identification}
}
\description{
Identify local extrema in a data frame
}
