% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.data.R
\name{load.data}
\alias{load.data}
\title{Load defined stocks historical data}
\usage{
load.data(stocks, from = "2020-08-01")
}
\arguments{
\item{stocks}{A vector of stock tickers}
}
\description{
Function loads historical data (open, high, low, close) for a given vector of stock tickers
}
