% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upside.maximizer.R
\name{upside.maximizer}
\alias{upside.maximizer}
\title{Identify minima in a data frame}
\usage{
upside.maximizer(
  stocks,
  from = "2020-08-01",
  span = 11,
  sell_upper = 0.2,
  sell_lower = 0.4,
  ...
)
}
\arguments{
\item{df}{A \code{data.frame} generated by the \code{load.data} function.}

\item{threshold}{An integer defining threshold for local extrema identification}
}
\value{
A list containing full data df, last local minimum and maximum records, upper and lower sell zone boundary
}
\description{
Identify local extrema in a data frame
}
